modified

+++ src/main/java/org/example/revhire/model/ResumeObjective.java
@@ -7,7 +7,7 @@ public class ResumeObjective {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
@@ -19,17 +19,17 @@ public class ResumeObjective {
     public ResumeObjective() {
     }
 
-    public ResumeObjective(Long id, User user, String objective) {
+    public ResumeObjective(Integer id, User user, String objective) {
         this.id = id;
         this.user = user;
         this.objective = objective;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
